todays adjustments

+++ priv/delete_uploads.exs
@@ -2,7 +2,10 @@ require Logger
 
 dir = Application.get_env(:erga, :uploads_directory)
 
-Logger.info("Deleting upload directory './#{dir}':")
+if dir do
+  # dir happends to be nil in some case wich would resoult in an error
 
-File.rm_rf!(dir)
-|> Enum.each(fn item -> Logger.info(item) end)
+  Logger.info("Deleting upload directory './#{dir}':")
+  File.rm_rf!(dir)
+  |> Enum.each(fn item -> Logger.info(item) end)
+end
